Added proper handling of old backups
Changed the naming convention of full backups to "${databases_identifier}-full"
and of incrementals to "${databases_identifier}-incremental". Also added removal
of these directories prior to the backup. If the cronjobs are properly
configured, there will always be one full backup and zero or more incremental
backups based on the latest full backup. The incremental backups are always
based on the last full backup, not the latest incremental backup, so even if an
external backup application (e.g. bacula or rsnapshot) fails to fetch an
incremental backup, restore from the full and latest fetched incremental will
still work properly.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>